Definition Report Updates:
- Added Definition Reports and allowed when selecting definition to be able to post the report for the definition

+++ src/definition-reports/services/definition-reports/definition-reports.service.ts
@@ -4,6 +4,7 @@ import { Repository } from 'typeorm';
 import { DefinitionReport } from '../../../typeorm/entities/definition-report';
 import { CreateDefinitionReportDto } from '../../dtos/create-definition-report.dto';
 import { UpdateDefinitionReportDto } from '../../dtos/update-definition-report.dto';
+import { User } from '../../../typeorm/entities/user';
 
 @Injectable()
 export class DefinitionReportsService {
@@ -13,51 +14,68 @@ export class DefinitionReportsService {
   ) {}
 
   async createReport(
-    dto: CreateDefinitionReportDto,
+    user: User,
+    reportDetails: CreateDefinitionReportDto,
   ): Promise<DefinitionReport> {
-    const report: DefinitionReport = new DefinitionReport();
-    Object.assign(report, dto);
-    return await this.definitionReportsRepository.save(report);
+    const newReport: DefinitionReport = this.definitionReportsRepository.create(
+      {
+        ...reportDetails,
+        userId: user.id,
+        createdAt: new Date(),
+      },
+    );
+    return await this.definitionReportsRepository.save(newReport);
   }
 
   async updateReport(
     id: number,
-    dto: UpdateDefinitionReportDto,
+    reportDetails: UpdateDefinitionReportDto,
   ): Promise<DefinitionReport> {
-    const updatedReport = await this.definitionReportsRepository.preload({
-      id,
-      ...dto,
-    });
-
-    if (!updatedReport) {
-      throw new NotFoundException(`Report with ID ${id} not found`);
-    }
-
-    return await this.definitionReportsRepository.save(updatedReport);
-  }
-
-  async deleteReport(id: number): Promise<void> {
-    const deleteResult = await this.definitionReportsRepository.delete(id);
-
-    if (!deleteResult.affected) {
-      throw new NotFoundException(`Report with ID ${id} not found`);
-    }
-  }
-
-  async getReportById(id: number): Promise<DefinitionReport> {
     const report: DefinitionReport =
       await this.definitionReportsRepository.findOne({
         where: { id },
       });
-
     if (!report) {
-      throw new NotFoundException(`Report with ID ${id} not found`);
+      throw new NotFoundException('Report not found');
     }
+    const updatedReport = await this.definitionReportsRepository.save({
+      ...report,
+      ...reportDetails,
+    });
+    return updatedReport;
+  }
 
+  async deleteReport(id: number) {
+    const report = await this.definitionReportsRepository.findOne({
+      where: { id },
+    });
+    if (!report) {
+      throw new NotFoundException('Report not found');
+    }
+    await this.definitionReportsRepository.delete(id);
+  }
+
+  async getReportById(id: number) {
+    const report = await this.definitionReportsRepository.findOne({
+      where: { id },
+    });
+    if (!report) {
+      throw new NotFoundException('Report not found');
+    }
     return report;
   }
 
-  async getAllReports(): Promise<DefinitionReport[]> {
+  async getAllReports() {
     return await this.definitionReportsRepository.find();
+  }
+
+  async getReportsByUserId(userId: number) {
+    return await this.definitionReportsRepository.find({ where: { userId } });
+  }
+
+  async getReportsByDefinitionId(definitionId: number) {
+    return await this.definitionReportsRepository.find({
+      where: { definitionId },
+    });
   }
 }
